Finish Chapter 5 + Reformat
Chapter 5 is now finished; it is now the day that they save the pink Butterfree in Chapter 6. I checked all the options for the convertion website so that it looks even cleaner (although it might not have even made a difference; who knows), and I've updated the CONTRIBUTING.md file with more info.