Issue fix implementation + grammar fixes + new Chapter 6
I've changed the story to reflect the issue fixes as well as performed another grammar check.
In addition, I've made a new chapter 6 (the old one is now chapter 7). This encompasses everything from after the dream to once they are going downstream; it's too much content to be a part of chapter 5 alone.
I have yet to do the scene description for chapter 6, and of course work still needs to be done on chapter 7.