Chapters 11 & 12 Scene Descriptions
As well as some other changes. It's not discernable in markup, but around line 782, Chapter 13's scene descriptions, I split the two scenes up (previously, as they were envisioned to be chapter descriptions, it was simply a slash separating them).
We're getting close to the end of the outline. Work work.