Finish Chapter 3
This chapter is pretty long, now that I think about it, compared to the other chapters. Although, this is apparently supposed to be the norm. I suppose we'll have to beef up the others.

Anyways, this'll need some serious proofreading in order to be accurate; heck, I might even purchase the movie for fact-checking. We'll see.

Onwards!