Chapter 4 - Part 1
I think this kind of wording is what we should strive for; full of details. Of course, this is just the draft; refining will come later.

I also changed up the wording in one of the spots in Chapter 3 to be more accurate to the plot.

The Pokémon Company released the movie on DVD. I'll try to pick that up so that I can do more plot fixes.