Chapter 4 - Part 2
This commit brings it up to when Charmander evolves into Charmeleon.

I had to improvise a bit with the battle itself. I don't think the movie itself shows more than a move or two, but you can't really make an entire battle out of one move, so I had to extend it. This is one of those things that'll need extra attention when revising.