Redo Chapter 3
Apologies that I didn't split this into a branch during writing (there would have been a lot of commits if I did); the thought just completely bypassed me until I had already started it and if I had already started it the commit content would be lopsided and it all just didn't work. So here's Chapter 3, redone as a part of #14!
(I made a branch so that spell-checking could be done)